Mr/upstream changes 2025 03 12

+++ apps/api/src/app/admin/admin.service.ts
@@ -1,7 +1,7 @@
-import { BenchmarkService } from '@ghostfolio/api/app/benchmark/benchmark.service';
 import { OrderService } from '@ghostfolio/api/app/order/order.service';
 import { SubscriptionService } from '@ghostfolio/api/app/subscription/subscription.service';
 import { environment } from '@ghostfolio/api/environments/environment';
+import { BenchmarkService } from '@ghostfolio/api/services/benchmark/benchmark.service';
 import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
 import { DataProviderService } from '@ghostfolio/api/services/data-provider/data-provider.service';
 import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service';
@@ -30,6 +30,7 @@ import {
   EnhancedSymbolProfile,
   Filter
 } from '@ghostfolio/common/interfaces';
+import { Sector } from '@ghostfolio/common/interfaces/sector.interface';
 import { MarketDataPreset } from '@ghostfolio/common/types';
 
 import { BadRequestException, Injectable, Logger } from '@nestjs/common';
@@ -108,7 +109,26 @@ export class AdminService {
     symbol
   }: AssetProfileIdentifier) {
     await this.marketDataService.deleteMany({ dataSource, symbol });
-    await this.symbolProfileService.delete({ dataSource, symbol });
+
+    const currency = getCurrencyFromSymbol(symbol);
+    const customCurrencies = (await this.propertyService.getByKey(
+      PROPERTY_CURRENCIES
+    )) as string[];
+
+    if (customCurrencies.includes(currency)) {
+      const updatedCustomCurrencies = customCurrencies.filter(
+        (customCurrency) => {
+          return customCurrency !== currency;
+        }
+      );
+
+      await this.putSetting(
+        PROPERTY_CURRENCIES,
+        JSON.stringify(updatedCustomCurrencies)
+      );
+    } else {
+      await this.symbolProfileService.delete({ dataSource, symbol });
+    }
   }
 
   public async get(): Promise<AdminData> {
@@ -260,6 +280,7 @@ export class AdminService {
             scraperConfiguration: true,
             sectors: true,
             symbol: true,
+            SymbolProfileOverrides: true,
             tags: true
           }
         }),
@@ -315,11 +336,10 @@ export class AdminService {
             Order,
             sectors,
             symbol,
+            SymbolProfileOverrides,
             tags
           }) => {
-            const countriesCount = countries
-              ? Object.keys(countries).length
-              : 0;
+            let countriesCount = countries ? Object.keys(countries).length : 0;
 
             const lastMarketPrice = lastMarketPriceMap.get(
               getAssetProfileIdentifier({ dataSource, symbol })
@@ -333,7 +353,34 @@ export class AdminService {
                 );
               })?._count ?? 0;
 
-            const sectorsCount = sectors ? Object.keys(sectors).length : 0;
+            let sectorsCount = sectors ? Object.keys(sectors).length : 0;
+
+            if (SymbolProfileOverrides) {
+              assetClass = SymbolProfileOverrides.assetClass ?? assetClass;
+              assetSubClass =
+                SymbolProfileOverrides.assetSubClass ?? assetSubClass;
+
+              if (
+                (
+                  SymbolProfileOverrides.countries as unknown as Prisma.JsonArray
+                )?.length > 0
+              ) {
+                countriesCount = (
+                  SymbolProfileOverrides.countries as unknown as Prisma.JsonArray
+                ).length;
+              }
+
+              name = SymbolProfileOverrides.name ?? name;
+
+              if (
+                (SymbolProfileOverrides.sectors as unknown as Sector[])
+                  ?.length > 0
+              ) {
+                sectorsCount = (
+                  SymbolProfileOverrides.sectors as unknown as Prisma.JsonArray
+                ).length;
+              }
+            }
 
             return {
               assetClass,
